feat: remove bouncy effect
When spamming the movement button after hitting an obstacle/wall, the player seems to have a "bouncing effect". The solution is to have a fixed update where the rotation and movement is based on the rigid body. Also, freeze the constraint of the rigid body on x and z axis